Avoid burst editing after a panic ends
Wait till the next cycle, be civil
Part of an effort to get rid of the idea of "just keep going" after a panic stop - a panic stop should be a hard stop and runs shouldnt continue after being half ended. Not sure why I did it that way in the first place, but whatever.